move utility fn
